requirements.txt
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
black==20.8b1
cached-property==1.5.2
celery==4.4.7
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coverage==5.5
cryptography==3.3.1
distro==1.5.0
Django==3.1.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
docker==4.4.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
Faker==8.0.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
jsonschema==3.2.0
kombu==4.6.11
Markdown==3.3.3
mixer==7.1.2
mypy-extensions==0.4.3
packaging==20.4
paramiko==2.7.2
pathspec==0.8.0
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
redis==3.2.0
regex==2020.10.15
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
text-unidecode==1.3
texttable==1.6.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.57.0
whitenoise==5.2.0
